Resolve VM offer publish bug & extend tests
There was an issue that the `azpc vm publish` command was not
using the right API endpoint due to the use of the prepare
request method. Until the manifest is correctly used, this method
needs to reference the JSON listing configuration, meaning that it
must be able to be passed into the CLI command so that the user
can configure which config to use.