Update VM offer authentication to enable deletion
To enable VM offer deletion, there is no Cloud Partner Portal API
endpoint, so we must use the Partner Center API endpoint, which works.
To reduce the duplication of code, it made sense to use the inherited
super Submission class' delete method, as functionally it is the same.
To be able to do this, the get_auth method in the VM class needed
updating to be able to save the state of the CPP API and PC API auth
tokens, to prevent the authentication API methods calling more than
required.
The show method has been removed from the VM class and now will also use
the inherited show method from the Submission class. This decision was
made because the product_id is set in the Submission.show() method, which
is required for deletion.