Tidy up integration tests to delete created offers
The integration tests required set up of a VM offer to validate that
it was behaving as expected. Now, we are also able to delete the offers
once the test has been completed.
Have moved the set up and tear down functions into a pytest fixture.
The methods have been modularised so that it can be used for the create
and delete integration tests.