Resolve bug related to offer ID and display name
There is a bug in the Partner Ingestion Swagger spec where external ids
are not an included attribute for an Azure Product.
This has led to the CLI not returning the offer ID in the `show` method
and only the offer display name.
This was leading to an issue where for the CLI to function as expected,
we needed the offer id (GUID) and the display name to be the same, which
is an incorrect assumption.
This is no longer an issue now that the Swagger spec has been updated.